add swagger

+++ server/src/main/java/com/ourstocks/jwtapp/rest/BlogRestControllerV1.java
@@ -10,6 +10,8 @@ import com.ourstocks.jwtapp.repository.UserRepository;
 import com.ourstocks.jwtapp.security.jwt.JwtUser;
 import com.ourstocks.jwtapp.service.PostService;
 import com.ourstocks.jwtapp.service.UserService;
+import io.swagger.annotations.Api;
+import io.swagger.annotations.ApiOperation;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -25,6 +27,8 @@ import java.util.stream.Collectors;
 
 @RestController
 @RequestMapping(value = "/api/v1/blog/")
+@Api(value = "Rest controller для роботи з розділом Блог")
+@CrossOrigin
 public class BlogRestControllerV1 {
     private final PostService postService;
     private final UserService userService;
@@ -43,6 +47,7 @@ public class BlogRestControllerV1 {
     }
 
     @RequestMapping(value = "", method = RequestMethod.GET,produces = MediaType.APPLICATION_JSON_VALUE)
+    @ApiOperation(value = "Отримання всіх постів з БД, незалежно від прав юзера")
     public ResponseEntity<List<PostDTO>> getAllPosts(){
         List<Post> posts = postService.getAllPosts();
         if(posts.isEmpty()){
@@ -53,6 +58,7 @@ public class BlogRestControllerV1 {
     }
 
     @RequestMapping(value = "add", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
+    @ApiOperation(value = "Додати новий пост")
     public ResponseEntity<Post> addPost(@AuthenticationPrincipal JwtUser jwtUser,
                                         @RequestBody @Valid RequestPostDTO postDTO){
         User user = userService.findByUsername(jwtUser.getUsername());
@@ -63,6 +69,7 @@ public class BlogRestControllerV1 {
     }
 
     @RequestMapping(value = "{id}", method = RequestMethod.GET,produces = MediaType.APPLICATION_JSON_VALUE)
+    @ApiOperation(value = "Отримати конкретний пост, але доступ є тільки у підтвердженого від автора підписника")
     public ResponseEntity<Post> getPostById(@PathVariable(name = "id") Long id,
                                             @AuthenticationPrincipal JwtUser jwtUser) {
         Post post = postService.findPostById(id);
@@ -82,11 +89,16 @@ public class BlogRestControllerV1 {
     }
 
     @RequestMapping(value = "{id}", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<Post> updatePost(@PathVariable(name = "id") Long id,
+    @ApiOperation(value = "Оновити дані посту, тільки якщо ти автор посту")
+    public ResponseEntity<Post> updatePost(@AuthenticationPrincipal JwtUser jwtUser,
+                                           @PathVariable(name = "id") Long id,
                                            @RequestBody @Valid RequestPostDTO postDTO){
         Post post = postService.findPostById(id);
         if(post == null){
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+        if(!jwtUser.getEmail().equals(post.getAuthor().getEmail())){
+            return new ResponseEntity<>(HttpStatus.LOCKED);
         }
         post = postDTO.toPost(post);
         postService.savePost(post);
@@ -94,9 +106,14 @@ public class BlogRestControllerV1 {
     }
 
     @RequestMapping(value = "{id}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<String> deletePost(@PathVariable(name = "id") Long id){
+    @ApiOperation(value = "Видалити пост, тільки якщо ти автор посту")
+    public ResponseEntity<String> deletePost(@AuthenticationPrincipal JwtUser jwtUser,
+                                             @PathVariable(name = "id") Long id){
         if(!postService.existsPostById(id)){
             return new ResponseEntity<>("Post not found",HttpStatus.BAD_REQUEST);
+        }
+        if(!jwtUser.getEmail().equals(postService.findPostById(id).getAuthor().getEmail())){
+            return new ResponseEntity<>(HttpStatus.LOCKED);
         }
         postService.deletePost(id);
         String message = "Post with id: " + id + " successfully deleted";
@@ -104,6 +121,7 @@ public class BlogRestControllerV1 {
     }
 
     @RequestMapping(value = "{id}/addcomment", method = RequestMethod.POST, produces = MediaType.APPLICATION_JSON_VALUE)
+    @ApiOperation(value = "Додати коментар до посту")
     public ResponseEntity<Post> addCommentToPost(@AuthenticationPrincipal JwtUser jwtUser,
                                                  @PathVariable(name = "id") Long id,
                                                  @RequestBody CommentDTO commentDTO){
@@ -123,10 +141,15 @@ public class BlogRestControllerV1 {
         return new ResponseEntity<>(post, HttpStatus.OK);
     }
     @RequestMapping(value = "{id}/{id_comment}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
-    public ResponseEntity<String> deleteCommentFromPost(@PathVariable(name = "id_comment") Long id_comment){
+    @ApiOperation(value = "Видалити коментар, тільки якщо ти автор коментаря")
+    public ResponseEntity<String> deleteCommentFromPost(@AuthenticationPrincipal JwtUser jwtUser,
+                                                        @PathVariable(name = "id_comment") Long id_comment){
         PostComment comment = postCommentRepository.findById(id_comment).orElse(null);
         if(comment == null) {
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
+        if(!jwtUser.getEmail().equals(comment.getAuthor().getEmail())){
+            return new ResponseEntity<>(HttpStatus.LOCKED);
         }
         String response = "Comment from post " + comment.getPost()
                 + " with id" + id_comment + " successfully deleted";
